Move providers into resources and use resource.action
