+++ lib/chef/resource/chef_data_bag.rb
@@ -1,13 +1,10 @@
 require 'cheffish'
-require 'chef_compat/resource'
+require 'cheffish/base_resource'
 
 class Chef
   class Resource
-    class ChefDataBag < ChefCompat::Resource
+    class ChefDataBag < Cheffish::BaseResource
       resource_name :chef_data_bag
-
-      allowed_actions :create, :delete, :nothing
-      default_action :create
 
       def initialize(*args)
         super
@@ -17,6 +14,50 @@ class Chef
       property :name, :kind_of => String, :regex => Cheffish::NAME_REGEX, :name_attribute => true
 
       property :chef_server, :kind_of => Hash
+
+
+      action :create do
+        if !current_resource_exists?
+          converge_by "create data bag #{new_resource.name} at #{rest.url}" do
+            rest.post("data", { 'name' => new_resource.name })
+          end
+        end
+      end
+
+      action :delete do
+        if current_resource_exists?
+          converge_by "delete data bag #{new_resource.name} at #{rest.url}" do
+            rest.delete("data/#{new_resource.name}")
+          end
+        end
+      end
+
+      action_class.class_eval do
+        def load_current_resource
+          begin
+            @current_resource = json_to_resource(rest.get("data/#{new_resource.name}"))
+          rescue Net::HTTPServerException => e
+            if e.response.code == "404"
+              @current_resource = not_found_resource
+            else
+              raise
+            end
+          end
+        end
+
+        #
+        # Helpers
+        #
+        # Gives us new_json, current_json, not_found_json, etc.
+
+        def resource_class
+          Chef::Resource::ChefDataBag
+        end
+
+        def json_to_resource(json)
+          Chef::Resource::ChefDataBag.new(json['name'], run_context)
+        end
+      end
     end
   end
 end
