+++ lib/cheffish/recipe_dsl.rb
@@ -24,20 +24,6 @@ require 'chef/resource/chef_role'
 require 'chef/resource/chef_user'
 require 'chef/resource/private_key'
 require 'chef/resource/public_key'
-require 'chef/provider/chef_acl'
-require 'chef/provider/chef_client'
-require 'chef/provider/chef_container'
-require 'chef/provider/chef_data_bag'
-require 'chef/provider/chef_data_bag_item'
-require 'chef/provider/chef_environment'
-require 'chef/provider/chef_group'
-require 'chef/provider/chef_mirror'
-require 'chef/provider/chef_node'
-require 'chef/provider/chef_organization'
-require 'chef/provider/chef_role'
-require 'chef/provider/chef_user'
-require 'chef/provider/private_key'
-require 'chef/provider/public_key'
 
 
 class Chef
