+++ spec/integration/chef_client_spec.rb
@@ -2,7 +2,6 @@ require 'support/spec_support'
 require 'cheffish/rspec/chef_run_support'
 require 'support/key_support'
 require 'chef/resource/chef_client'
-require 'chef/provider/chef_client'
 
 repo_path = Dir.mktmpdir('chef_repo')
