+++ spec/integration/chef_container_spec.rb
@@ -1,7 +1,5 @@
 require 'support/spec_support'
 require 'cheffish/rspec/chef_run_support'
-require 'chef/resource/chef_container'
-require 'chef/provider/chef_container'
 
 describe Chef::Resource::ChefContainer do
   extend Cheffish::RSpec::ChefRunSupport
