+++ spec/integration/chef_node_spec.rb
@@ -1,7 +1,5 @@
 require 'support/spec_support'
 require 'cheffish/rspec/chef_run_support'
-require 'chef/resource/chef_node'
-require 'chef/provider/chef_node'
 
 describe Chef::Resource::ChefNode do
   extend Cheffish::RSpec::ChefRunSupport
