+++ spec/integration/chef_organization_spec.rb
@@ -1,7 +1,5 @@
 require 'support/spec_support'
 require 'cheffish/rspec/chef_run_support'
-require 'chef/resource/chef_organization'
-require 'chef/provider/chef_organization'
 
 describe Chef::Resource::ChefOrganization do
   extend Cheffish::RSpec::ChefRunSupport
