+++ spec/integration/chef_user_spec.rb
@@ -1,8 +1,6 @@
 require 'support/spec_support'
 require 'cheffish/rspec/chef_run_support'
 require 'support/key_support'
-require 'chef/resource/chef_user'
-require 'chef/provider/chef_user'
 
 repo_path = Dir.mktmpdir('chef_repo')
