+++ spec/integration/private_key_spec.rb
@@ -1,9 +1,5 @@
 require 'support/spec_support'
 require 'cheffish/rspec/chef_run_support'
-require 'chef/resource/private_key'
-require 'chef/provider/private_key'
-require 'chef/resource/public_key'
-require 'chef/provider/public_key'
 require 'support/key_support'
 
 repo_path = Dir.mktmpdir('chef_repo')
