+++ spec/integration/recipe_dsl_spec.rb
@@ -1,7 +1,5 @@
 require 'support/spec_support'
 require 'cheffish/rspec/chef_run_support'
-require 'chef/resource/chef_node'
-require 'chef/provider/chef_node'
 require 'tmpdir'
 
 describe 'Cheffish Recipe DSL' do
